Add configurable Maps API Key to clients
The Google Maps API Key is no longer hard coded into the clients.
It can now be set like other config variables, via a "MAPS_API_KEY"
environment variable, or using a "config.json". Unlike other config
variables, if not set, there is no default. Instead, the first
user to browse to a map page will be prompted to enter an API key
which will be saved to config.json.

Signed-off-by: ZacX Delventhal <zac@bitwise.io>