Add comprehensive shell container
Includes all dependencies to run all scripts and commands in the Supply Chain
project. This container is used in the compose file to run protogen and
build client files, removing the need for users or the Server container to
run these commands.

Also includes updates to the Readme that cover the use of the shell, and
defaults to Server to serving both clients.

Signed-off-by: Zac Delventhal <zac@bitwise.io>