Seed example data on docker-compose up
This updates the server scripts to retry the server if unavailable,
and takes advantage of this behavior to add the data seeding scripts
to the command for the shell container. This means that `docker-compose up`
will now start up all components AND seed the web apps with both the
requisite RecordTypes and some sample data.

Signed-off-by: Zac Delventhal <zac@bitwise.io>