Update Protobuf data types to make room for expansion
This changes the indexes for the PropertySchema.DataType enum,
as well as the various `<type>_value` properties. This will not
functionally change anything, but makes room further new planned
data types.

Note that INT will eventually be replaced in place with NUMBER, and
FLOAT will be removed.

Also adds an UNSET key at 0 to reflect current Sawtooth best practices.

Signed-off-by: Zac Delventhal <zac@bitwise.io>