Remove Python transaction processor
As the Rust TP replicates all of the functionality of the Python TP,
it is no longer useful, and unecessarily complicates development to
maintain. This removes the TP as well as a number of scripts specific
to Python development, and moves the Rust TP up to the root level of
the processor directory.

Note that since the integration tests are written in Python, a few
supporting scripts like bin/protogen remain. Those should be refactored
out whenever possible.

Signed-off-by: Zac Delventhal <zac@bitwise.io>