Implement retry when failing to connect to RethinkDB
The Ledger-sync and Server are designed to wait for the
proper database and table to become available before
starting up. However, this process fails with a connection
error if they query for the table before Rethink starts up.

This has not been a poblem in the past as Rethink is typically
quicker to start up than either process, but has recently
caused errors for some users. With this update, both processes
will retry their Rethink connection if it fails at first, in
addition to querying for the proper table to be available.

Signed-off-by: Zac Delventhal <zac@bitwise.io>