Add wait time notification to test logs
The start time for the Rust TP is unpredictable and can be quite long,
so tests wait 60 seconds to start. This can give the false impression
that the tests have failed silently. A log has been added to help
clarify. The `command` for the test container has also be cleaned up.

Signed-off-by: Zac Delventhal <zac@bitwise.io>