Add ENUM data type to transaction processor
Adds a new ENUM data type both to the protobuf property definitions,
and to the TP's handler. This will allow properties to be created
with a limited set of string values, stored as integers. These valid
`enum_options` are defined first in the PropertySchema, and later copied
to the Property, where they are stored and referenced during updates.

Signed-off-by: Zac Delventhal <zac@bitwise.io>