Add ENUM data type to Server's REST API
In order to return the string name of enums rather than the index store
on-chain, the Property for a PropertyPage must be fetched to reference
the enum_options. This is done in the ledger sync to match the behavior
of the DataType enums which are converted to strings befor saving to
the database.

Signed-off-by: Zac Delventhal <zac@bitwise.io>