Update core components to use NUMBER data type
Affects the TP, server, and core Protobufs. Removes the FLOAT data type,
and renames INT to NUMBER. Also adds the `number_exponent` field, which
will be used to annotate numbers to indicate the integer value
represents a fractional unit.

This change ensures that Supply Chain developers follow best practices by
only used fixed point data to represent numbers.

Because this is a breaking change, the FAMILY_VERSION has been incremented
to 1.1 in the TP and server.

Signed-off-by: Zac Delventhal <zac@bitwise.io>