Update clients to use NUMBER instead of INT and FLOAT
This updates the fish and asset clients to be compatible with the new
NUMBER data type, which replaces both INT and FLOAT. It does _not_ take
advantage of the number_exponent property to dynamically turn
numbers into decimal notation, and instead relies on the existing
hard coded solution. Later clients will take advantage of this
new functionality, but the existing clients are getting minimal
updates.

Als increments the FAMILY_VERSION used by the clients to 1.1.

Signed-off-by: Zac Delventhal <zac@bitwise.io>